+++ apps/ministreaming/java/src/net/i2p/client/streaming/I2PSocketEepGet.java
@@ -167,12 +167,12 @@ public class I2PSocketEepGet extends EepGet {
                                 byte[] b = Base32.decode(host.substring(0, 52));
                                 if (b != null) {
                                     Hash h = Hash.create(b);
-                                    dest = sess.lookupDest(h, 30*1000);
+                                    dest = sess.lookupDest(h, (long) 30*1000);
                                 } else {
                                     dest = null;
                                 }
                             } else {
-                                dest = sess.lookupDest(host, 30*1000);
+                                dest = sess.lookupDest(host, (long) 30*1000);
                             }
                         } catch (I2PSessionException ise) {
                             dest = null;
